requirements.txt
python-dotenv
requests
pandas
snowflake-connector-python[pandas]
dbt-snowflake